# Perf improvement for 2 arg show for `UnitRange` type